requirements.txt
pyyaml
snowflake-connector-python
pandas
networkx
pyvis
streamlit